requirements_cleaned.txt
# Core Django and API
Django==4.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-ratelimit==4.1.0
django-redis==5.4.0
django-simple-history==3.4.0

# Database
psycopg==3.2.3
psycopg-binary==3.2.3
dj-database-url==2.1.0

# Task Queue
celery==5.3.6
redis==5.0.1
amqp==5.3.1
billiard==4.2.2
kombu==5.6.0
vine==5.1.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0

# Authentication & Security
Authlib==1.6.5
PyJWT==2.8.0
cryptography==41.0.7

# AI/ML and LangChain
openai==1.68.2
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.63
langchain-openai==0.3.11
langchain-pinecone==0.2.13
langchain-text-splitters==0.3.8
langgraph==0.2.50
langgraph-checkpoint==2.1.2
langgraph-sdk==0.1.74
langsmith==0.1.147
together==1.2.1
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
tiktoken==0.8.0

# Vector Database
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.4
pdfminer.six==20231228
pypdfium2==5.0.0

# HTTP Clients
requests==2.32.5
requests-toolbelt==1.0.0
httpx==0.28.1
httpx-sse==0.4.3
aiohttp==3.13.2
aiohttp-retry==2.9.1
httpcore==1.0.9
httplib2==0.31.1

# Messaging
twilio==9.0.4

# Utilities
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.2.1
pytz==2024.2
tzdata==2025.2
tzlocal==5.3.1
dateparser==1.2.2
chardet==5.2.0
charset-normalizer==3.4.4

# Data Processing
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
marshmallow==3.26.1
dataclasses-json==0.6.7
orjson==3.11.4
ormsgpack==1.12.1

# Monitoring & Logging
sentry-sdk==1.40.6

# Testing (keep for development)
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==4.1.0
pytest-django==4.7.0
factory-boy==3.3.0
Faker==37.12.0
hypothesis==6.92.1
coverage==7.11.3

# Development Tools (keep for development)
black==23.12.1
flake8==7.0.0
isort==5.13.2
pre-commit==3.5.0

# Core Python Dependencies
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
pycparser==2.23
idna==3.11
urllib3==2.5.0
six==1.17.0
setuptools
wheel

# Additional required dependencies
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cachetools==6.2.2
cfgv==3.5.0
click==8.3.0
distlib==0.4.0
distro==1.9.0
eval_type_backport==0.2.2
filelock==3.20.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
identify==2.6.15
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.19.2
pyparsing==3.2.5
pypng==0.20220715.0
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
rpds-py==0.28.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
shellingham==1.5.4
simsimd==6.5.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
sqlparse==0.5.3
stevedore==5.6.0
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
virtualenv==20.35.4
wcwidth==0.2.14
yarl==1.22.0